Vulkan: Match desctriptor pool sizes to descriptor set layouts
When a descriptor pool is created, a list of descriptor types and counts
are given to vkCreateDescriptorPool(). Later, when allocating a
descriptor set from the pool, we pass along a descriptor set layout to
vkAllocateDescriptorSets() which is used to determine how many of each
type of descriptor (i.e., binding) to allocate from the pool.

In order for our "free descriptor set" counts to be accurate for each
pool, the descriptor pools need to be created with descriptor counts
that match the descriptor set layout binding counts.

This change fixes a bug where the descriptor set layouts were created
with more bindings than the descriptor pool sizes, causing the "free
desctriptor set" count to be inaccurate, leading to allocating too many
descriptor sets from a pool.

Bug: angleproject:3570
Test: CQ with kDefaultDescriptorPoolMaxSets = 1
Test: CQ

+++ src/libANGLE/renderer/vulkan/vk_helpers.h
@@ -287,7 +287,7 @@ class DynamicDescriptorPool final : angle::NonCopyable
     // an individual set.  The pool size will be calculated accordingly.
     angle::Result init(ContextVk *contextVk,
                        const VkDescriptorPoolSize *setSizes,
-                       uint32_t setSizeCount);
+                       size_t setSizeCount);
     void destroy(VkDevice device);
     void release(ContextVk *contextVk);
